Do not use subpackages for ROFL repositories
Autotools subpackages would be the natural solution for this, but unfortunately
they forgot that in most cases, and  dependencies is a good example,
the configure options for the inner subpackages will in general be different
than the ones in the main package, and likely be dependent on them.

So we define them as regular Makefile targets. The `compile_autotools_pkg.sh` is I
guess not needed, but I don't have the time to figure out how to embed it
into the Makefile target without automake screwing it.