Fixed pass configure opts to rofl-pipeline
Prior to this commit, the pipeline optimizations
e.g. --with-pipeline-lockless were always passed to the library
if the driver was supporting it. There was no way to disable
them.

This commit fixes this issue. During configure (xdpd) time,
it checks whether the user is overriding the defaults for
rofl-pipeline, and compiles rofl-datapath accordingly.

Reported-by: Tobias Jungel <tobias.jungel@bisdn.de>