gnu-linux: temporally work-around veth NS issues
As discussed in #44, ioport_mmap attempts to disable all
offloads both in the interface and, in case the interface is a
veth pair, the other edge.

However, if the other edge is running in another namespace, the
calls will fail and the port initialization will completely fail
as it is not scheduled in the I/O due to the bring up failure.

This commit works-around this issue by warning the user that
offloads cannot be disabled but continuing with the port UP
command and I/O scheduling.

This is a temporary work around that forces users to manually
disable offloads on BOTH edges. This should be properly fixed by
detecting in which namespace the pair is, and moving temporally
there to disable the offloads, so freeing the user to deal with
these low level details.